requirements.txt
requests~=2.26.0
python-dotenv~=0.19.0
pandas~=1.3.2
requests_oauthlib
openpyxl
xlsxWriter